+++ lib/meta/models.dart
@@ -25,8 +25,28 @@ class Model extends ConceptEntity<Model> {
     selectedElements.forEach((concept) => entryList.add(concept));
     return entryList;
   }
+  
+  // ordered by external parent count (from 0 to ...)
+  List<Concept> get orderedEntryConcepts {
+    List<Concept> entryConcepts = this.entryConcepts;
+    var orderedEntryConceptsCount = 0;
+    List<Concept> orderedEntryConcepts = new List<Concept>();
+    for (var c = 0; c < 10; c++) {
+      for (Concept entryConcept in entryConcepts) {
+        var count = entryConcept.parents.externalCount;
+        if (count == c) {
+          orderedEntryConcepts.add(entryConcept);
+          if (++orderedEntryConceptsCount == entryConcepts.length) {
+            return orderedEntryConcepts;
+          }
+        }
+      }
+    }
+    return orderedEntryConcepts;
+  }
 
   int get entryConceptCount => entryConcepts.length;
+  int get orderedEntryConceptCount => orderedEntryConcepts.length;
 
   Concept getEntryConcept(String entryConceptCode) {
     Concept concept = concepts.singleWhereCode(entryConceptCode);