+++ CHANGELOG.md
@@ -4,6 +4,8 @@ based on [Semantic Versioning](http://semver.org/)
 
 *2.0.0* 2014-03-09
 
++ concept external parents and external requitred parents
++ model ordered (by external parent count) entry concepts
 + generate model init with initChildren ordered by external parent count
 + external parent count for concept parents
 + rename base type to origin type